refactor: move files to new folders structure
