+++ app/globals.scss
@@ -1,0 +1,359 @@
+:root {
+  font-family: system-ui, Avenir, Helvetica, Arial, sans-serif;
+  line-height: 1.5;
+  font-weight: 400;
+
+  color-scheme: light dark;
+
+  font-synthesis: none;
+  text-rendering: optimizeLegibility;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
+
+[data-theme='light'] {
+  --bg-color: #ffffff;
+  --text-color: #5998ff;
+  --primary-color: #2563eb;
+}
+
+[data-theme='dark'] {
+  --bg-color: #121212;
+  --text-color: #f0f0f0;
+  --primary-color: #3b82f6;
+}
+
+body {
+  background-color: var(--bg-color);
+  color: var(--text-color);
+  transition:
+    background-color 0.3s,
+    color 0.3s;
+}
+
+.theme-toggle button {
+  background: none;
+  border: none;
+  cursor: pointer;
+  font-size: 1.5rem;
+}
+
+html {
+  scroll-behavior: smooth;
+}
+
+* {
+  box-sizing: border-box;
+  margin: 0;
+
+  text-align: center;
+
+  transition:
+    background-color 0.6s ease-out,
+    color 0.1s ease-out,
+    all 0.5s ease;
+
+  &::before {
+    box-sizing: border-box;
+  }
+
+  &::after {
+    box-sizing: border-box;
+  }
+
+  a {
+    font-weight: 500;
+    color: #646cff;
+    text-decoration: none;
+
+    &:hover {
+      color: #535bf2;
+    }
+  }
+}
+
+h1 {
+  font-size: 2.2em;
+}
+
+button,
+.button {
+  border-radius: 8px;
+  border: 1px solid transparent;
+  padding: 0.5em 1em;
+  font-size: 1em;
+  font-weight: 500;
+  font-family: inherit;
+  background-color: #1a1a1a;
+  cursor: pointer;
+  transition: border-color 0.25s;
+
+  &:hover {
+    border-color: #646cff;
+  }
+
+  &:focus,
+  :focus-visible {
+    outline: 4px auto -webkit-focus-ring-color;
+  }
+}
+
+@media (prefers-color-scheme: light) {
+  :root {
+    color: #8dc9ff;
+    background-color: #ffffff;
+  }
+  a:hover {
+    color: #747bff;
+  }
+  button {
+    background-color: #f9f9f9;
+  }
+  .header {
+    background-color: #0b0036;
+  }
+}
+
+.header {
+  background-color: #312a4d;
+  padding: 1em;
+  border-radius: 0 0 1em 1em;
+  width: 100%;
+  display: flex;
+  align-content: center;
+  align-items: center;
+  justify-content: space-around;
+  flex-wrap: wrap;
+  gap: 1rem;
+}
+
+.main {
+  position: relative;
+  padding-top: 1.5rem;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  gap: 1rem;
+}
+
+.card-list {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+  gap: 2rem;
+}
+
+.error {
+  color: darkred;
+  text-transform: capitalize;
+  text-decoration: underline;
+}
+
+.padding {
+  padding: 2rem;
+}
+
+.about {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 1em;
+}
+
+.selection-items {
+  position: absolute;
+  top: 0;
+  left: 50%;
+  transform: translateX(-50%);
+  line-height: 1rem;
+  font-size: 1rem;
+}
+
+.search {
+  &-header {
+    min-height: 2.7rem;
+    width: 100%;
+    display: flex;
+    gap: 1rem;
+  }
+
+  &-container {
+    display: flex;
+    justify-content: flex-end;
+    gap: 0.5rem;
+    width: 50%;
+    padding-left: 1rem;
+    border-left: 1px solid var(--primary-color);
+  }
+
+  &-selection {
+    display: flex;
+    justify-content: flex-start;
+    width: 50%;
+    gap: 0.5rem;
+  }
+
+  &-body {
+    display: flex;
+    gap: 1rem;
+    align-items: flex-start;
+  }
+
+  &-panel {
+    flex: 2;
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+  }
+}
+
+.pagination {
+  display: flex;
+  gap: 1rem;
+  justify-content: center;
+  align-items: center;
+}
+
+.detail {
+  &-panel {
+    width: 100%;
+    flex: 1;
+    border-left: 1px solid #ddd;
+    padding-left: 1rem;
+    min-width: 250px;
+  }
+
+  &-container {
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+
+    img {
+      border-radius: 1rem;
+    }
+  }
+}
+
+.card {
+  position: relative;
+  width: 9rem;
+  height: 9rem;
+  border-radius: 20px;
+  overflow: hidden;
+  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.3);
+  cursor: pointer;
+
+  &-shadow {
+    background-size: contain;
+    background-position: center;
+    filter: blur(2px);
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    transform: scale(1.1);
+    z-index: 0;
+  }
+
+  &-container {
+    position: relative;
+    height: 100%;
+    padding: 1rem 1rem 2rem;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    overflow: hidden;
+    z-index: 2;
+  }
+
+  &-checkbox {
+    position: absolute;
+    bottom: 0.7rem;
+    left: 50%;
+    transform: translateX(-50%);
+    width: 9rem;
+    height: 1rem;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    gap: 0.5rem;
+    font-size: 0.7rem;
+    z-index: 3;
+  }
+
+  h2 {
+    font-size: 0.8rem;
+    line-height: 0.8rem;
+  }
+
+  p {
+    font-size: 0.5rem;
+  }
+
+  &:hover {
+    scale: 1.1;
+  }
+}
+
+@media (max-width: 600px) {
+  .search {
+    &-header {
+      width: 100%;
+      display: flex;
+      flex-direction: column-reverse;
+      gap: 1rem;
+
+      & > div {
+        width: 100%;
+        justify-content: space-around;
+      }
+    }
+
+    &-container {
+      padding-left: unset;
+      border-left: unset;
+    }
+
+    &-selection {
+      padding-top: 1rem;
+      border-top: 1px solid var(--primary-color);
+    }
+
+    &-body {
+      flex-direction: column;
+    }
+  }
+  .detail {
+    &-panel {
+      border-left: none;
+      border-top: 1px solid #ddd;
+      padding: 1rem;
+    }
+
+    &-container {
+      max-width: 100%;
+
+      img {
+        max-width: 100%;
+      }
+    }
+  }
+  .card {
+    border-radius: 0.5rem;
+    width: 8rem;
+    height: 10rem;
+    &:hover {
+      scale: 1.02;
+    }
+    &-list {
+      gap: 0.5rem;
+    }
+  }
+}
+.theme-select.button {
+  height: 100%;
+  padding: unset;
+}
