+++ components/Container/index.tsx
@@ -1,0 +1,8 @@
+'use client';
+
+import { type ReactNode } from 'react';
+import styles from './container.module.scss';
+
+export default function Container({ children }: { children: ReactNode }) {
+  return <div className={styles.container}>{children}</div>;
+}
