+++ components/Search/SearchButton/index.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+const SearchButtonText = 'Search';
+
+export default function SearchButton({ onClick }: { onClick: VoidFunction }) {
+  return <button onClick={onClick}>{SearchButtonText}</button>;
+}
