+++ helpers/isEmpty/index.ts
