--- src/components/Card/card.scss
@@ -1,61 +1,0 @@
-.card {
-  position: relative;
-  width: 9rem;
-  height: 9rem;
-  border-radius: 20px;
-  overflow: hidden;
-  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.3);
-  cursor: pointer;
-
-  &-shadow {
-    background-size: contain;
-    background-position: center;
-    filter: blur(2px);
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    transform: scale(1.1);
-    z-index: 0;
-  }
-
-  &-container {
-    position: relative;
-    height: 100%;
-    padding: 1rem 1rem 2rem;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    overflow: hidden;
-    z-index: 2;
-  }
-
-  &-checkbox {
-    position: absolute;
-    bottom: 0.7rem;
-    left: 50%;
-    transform: translateX(-50%);
-    width: 9rem;
-    height: 1rem;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    gap: 0.5rem;
-    font-size: 0.7rem;
-    z-index: 3;
-  }
-
-  h2 {
-    font-size: 0.8rem;
-    line-height: 0.8rem;
-  }
-
-  p {
-    font-size: 0.5rem;
-  }
-
-  &:hover {
-    scale: 1.1;
-  }
-}
