--- src/components/Navigation/index.tsx
@@ -1,13 +1,0 @@
-import { NavLink } from 'react-router';
-
-export default function Navigation() {
-  return (
-    <nav>
-      <NavLink to="search?page=1" end>
-        Home
-      </NavLink>
-      {' | '}
-      <NavLink to="about">About</NavLink>
-    </nav>
-  );
-}
