--- src/components/Search/SearchButton/index.tsx
@@ -1,5 +1,0 @@
-const SearchButtonText = 'Search';
-
-export default function SearchButton({ onClick }: { onClick: VoidFunction }) {
-  return <button onClick={onClick}>{SearchButtonText}</button>;
-}
