--- src/components/features/NotFound/index.tsx
@@ -1,10 +1,0 @@
-import { NavLink } from 'react-router';
-
-export default function NotFound() {
-  return (
-    <>
-      <h1>Page Not Found</h1>
-      <NavLink to={'/'}>Go Back!</NavLink>
-    </>
-  );
-}
