--- src/main.tsx
@@ -1,17 +1,0 @@
-import { createRoot } from 'react-dom/client';
-import App from './App.tsx';
-import ErrorBoundary from './components/error/ErrorBoundary';
-import { ThemeProvider } from './context/ThemeContext.tsx';
-
-const rootElement: HTMLDivElement = document.createElement('div');
-rootElement.id = 'root';
-rootElement.dataset.testid = 'root-element';
-document.body.append(rootElement);
-
-createRoot(rootElement).render(
-  <ErrorBoundary>
-    <ThemeProvider>
-      <App />
-    </ThemeProvider>
-  </ErrorBoundary>
-);
