+++ store/selectedItemsStore.ts
